Cleaned inventory components
